requirements.txt
requests
kafka-python
PyYAML
Flask
